+++ source/Manga.Infrastructure/EntityFrameworkDataAccess/Migrations/MangaContextModelSnapshot.cs
@@ -45,7 +45,7 @@ namespace Manga.Infrastructure.EntityFrameworkDataAccess.Migrations
 
                 b.Property<Guid>("AccountId");
 
-                b.Property<double?>("Amount");
+                b.Property<decimal?>("Amount");
 
                 b.Property<DateTime>("TransactionDate");
 
@@ -70,7 +70,7 @@ namespace Manga.Infrastructure.EntityFrameworkDataAccess.Migrations
 
                 b.Property<Guid>("AccountId");
 
-                b.Property<double?>("Amount");
+                b.Property<decimal?>("Amount");
 
                 b.Property<DateTime>("TransactionDate");
