+++ tests/Manga.UnitTests/UseCaseTests/Transfer/PositiveDataSetup.cs
@@ -2,7 +2,7 @@ namespace Manga.UnitTests.UseCaseTests.Transfer
 {
     using Xunit;
 
-    internal sealed class PositiveDataSetup : TheoryData<double, double>
+    internal sealed class PositiveDataSetup : TheoryData<decimal, decimal>
     {
         public PositiveDataSetup()
         {
